Fixes type of port parameter in cfg files
The port parameter was declared as an int_t in the cfg files but in the
code it is expected to be a string. When the dynamic_reconfigure server
is started, the data type on the parameter server was changed from
string to int. This resulted in getParam of the release 1.2.0 to return
False effectively killing any respawn functionality.