Implement a watchdog that monitors whether the scanner is accessible or not
Our current workflow with the EAE involves powercycling the base truck
(by turning the key switch on/off) during bringup restarts or just after them.

This causes two main problems in the following scenarios.

Scenario 1:
- user stops bringup
- turns the key switch off
- starts the bringup (scanner is off at this point)
  r2000_node tries to connect with it, but fails.
  It terminates and respawns repatedly.
  At some point, bringup gaves up on it and it doesn't even respawn.

- user turns on the key switch On

Now, the scanner is back but r2000_node will not respawn.

Scenario 2:
- user turns the key switch off
- restarts the bringup
- r2000_node is trying to do clean up.
  but it cannot join its reader & writer threads when the scanner is off.
  these threads linger between bringup restarts.
- user turns the key switch on

This PR adds a somewhat hacky watchdog that monitors whether the scanner is
still accessible or not.

Closes: SW-82782